Create main_script
This file gives very important insights into the data logged from OPC N2 Sensor, The original code was modified to make it run in a loop. The current code prints the data continuously as well as logs the information in CSV format. A very important change is that instead of bin numbers as the header of the file, this code gives the actual bin sizes as heading, this will help the researchers to know the exact size of the particle and its concentration rather than bin numbers.